Extents: Split out extent struct.
